requirements.txt
numpy
pandas
PyWavelets
scipy
biopython
matplotlib
